Accessing a section's full text, HTML tag fixes
This commit adds a #text method to the Section class. The method returns
the full text in the section. Without this method, calling #text on a
section would return all of the text on the page.

This PR also addresses minor issues with closing tags in the test site.
HTMl tags were appearing when `page.text` was called on the test site
due to several missing tags.