Alter where the block evaluation is done for #when_loaded
This is technically a bug, because load validations were not being performed when running #load unless you provided a block. Even if that block was spurious i.e. { puts :foo }

By moving the check (And passing the block down the chain as before), we can still perform the load validations, and only yield the block if one is provided